fix publish script

+++ make_package_from_git.sh
@@ -1,10 +1,12 @@
 #!/usr/bin/env bash
 rm -rf aided_reading
 git clone https://github.com/mechatroner/aided_reading.git
-previous_version=$( cat words_discoverer_chrome_ext/manifest.json | grep '"version":' | cut -d '"' -f 4 )
-if [ -n "$previous_version" ] ; then
-    rm -rf "words_discoverer_chrome_ext.v$previous_version"
-    mv words_discoverer_chrome_ext "words_discoverer_chrome_ext.v$previous_version"
+if [ -d words_discoverer_chrome_ext ] ; then
+    previous_version=$( cat words_discoverer_chrome_ext/manifest.json | grep '"version":' | cut -d '"' -f 4 )
+    if [ -n "$previous_version" ] ; then
+        rm -rf "words_discoverer_chrome_ext.v$previous_version"
+        mv words_discoverer_chrome_ext "words_discoverer_chrome_ext.v$previous_version"
+    fi
 fi
 cp -r aided_reading/words_discoverer_chrome words_discoverer_chrome_ext
 rm words_discoverer_chrome_ext.zip
